requirements.txt
tensorflow>=2.8.0,<3.0
numpy>=1.22.0,<2.0
pygame>=2.1.0,<3.0
Pillow>=9.0.0,<10.0
argparse==1.4.0
